Rename Parallel/Algorithm.hpp to Parallel/DistributedObject.hpp
