+++ src/Parallel/CMakeLists.txt
@@ -21,7 +21,6 @@ spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
-  Algorithm.hpp
   AlgorithmMetafunctions.hpp
   ArrayIndex.hpp
   Callback.hpp
@@ -29,6 +28,7 @@ spectre_target_headers(
   CharmPupable.hpp
   CharmRegistration.hpp
   CreateFromOptions.hpp
+  DistributedObject.hpp
   GetSection.hpp
   GlobalCache.hpp
   GlobalCacheDeclare.hpp
